+++ src/caselawclient/types.py
@@ -1,5 +1,4 @@
 from dataclasses import dataclass, field
-from typing import NamedTuple
 
 
 @dataclass
@@ -70,9 +69,40 @@ class DocumentIdentifierValue(str):
     pass
 
 
-SuccessFailureMessageTuple = NamedTuple("SuccessFailureMessageTuple", [("success", bool), ("messages", list[str])])
-"""
-A tuple used to return if an operation has succeeded or failed (and optionally a list of messages associated with that operation).
+class SuccessFailureMessageTuple(tuple[bool, list[str]]):
+    """
+    Return whether an operation has succeeded or failed
+    (and optionally a list of messages associated with that operation).
+    Typically the messages will be exposed to the end-user.
+    Use only where a failure is a routine event (such as during validation).
+    """
 
-This should only be used where a failure is considered a routine part of the application (eg during validation options); where an unexpected action has led to a failure the application should raise an appropriate exception.
-"""
+    def __new__(cls, success: bool, messages: list[str]) -> "SuccessFailureMessageTuple":
+        return super().__new__(cls, [success, messages])
+
+    @property
+    def success(self) -> bool:
+        return self[0]
+
+    @property
+    def messages(self) -> list[str]:
+        return self[1]
+
+    def __repr__(self) -> str:
+        return f"SuccessFailureMessageTuple({self.success!r}, {self.messages!r})"
+
+    def __bool__(self) -> bool:
+        return self.success
+
+    def __or__(self, other: "SuccessFailureMessageTuple") -> "SuccessFailureMessageTuple":
+        """Allows us to write combined_tuple = first_tuple | second_tuple"""
+        return SuccessFailureMessageTuple(self.success and other.success, self.messages + other.messages)
+
+
+def SuccessTuple() -> SuccessFailureMessageTuple:
+    return SuccessFailureMessageTuple(True, [])
+
+
+def FailureTuple(message: str | list[str]) -> SuccessFailureMessageTuple:
+    messages = message if isinstance(message, list) else [message]
+    return SuccessFailureMessageTuple(success=False, messages=messages)