+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog 1.0.0].
 
+## Unreleased
+
+### Feat
+
+- **types** SuccessFailureMessageTuple acts as a boolean and keeps previous behaviour
+
 ## v42.0.0 (2025-10-24)
 
 ### BREAKING CHANGE